django_api/sqlite_to_postgres/requirements.txt
flake8
python-dotenv
sqlite3
psycopg2
dataclasses


